requirements.txt
fastapi
mlflow==1.30.0
uvicorn
psycopg2-binary